+++ socket/src/decorators/index.ts
@@ -1,3 +1,4 @@
-export { Namespace, ServerMiddleware, GlobalMiddleware, Middleware } from './class';
-export { Connection, Disconnect, GlobalEvent, Event } from './property';
+export { Namespace } from './ns';
+export { ServerMiddleware, GlobalMiddleware, Middleware } from './middleware';
+export { Connection, Disconnect, GlobalEvent, Event } from './events';
 export { IO, Socket, Args, Callback } from './param';