+++ socket/src/socket.ts
@@ -1,12 +1,16 @@
 import {
   ParameterType,
   Listener,
-  SocketIOMeta,
+  Meta,
   Injectable,
   SocketIOClass,
-  ParameterConfiguration,
-  Params
+  Param,
+  DecoratorsArtifacts,
+  DEFAULT_NAMESPACE,
+  MiddlewareType,
+  Middleware
 } from './interface';
+import { getMeta } from './meta';
 
 /**
  * Dummy function to ensure, that callback exists
@@ -19,7 +23,7 @@ function noop() {}
  * @param {SocketIO.Socket} socket
  * @returns {SocketIO.Socket}
  */
-function getSocket(item: ParameterConfiguration, socket: SocketIO.Socket) {
+function getSocket(item: Param, socket: SocketIO.Socket) {
   return item.data ? new item.data(socket) : socket;
 }
 
@@ -29,15 +33,18 @@ function getSocket(item: ParameterConfiguration, socket: SocketIO.Socket) {
  * @param socket
  * @param params Meta parameters
  * @param eventArgs incoming arguments
- * @returns {Array} returns arguments array, or io and socket, event args (default) by default
+ * @returns {Array} returns arguments array, or io and socket, event args by default
  */
-function extractParameters(
+function extractArguments(
   io: SocketIO.Server | SocketIO.Namespace,
   socket: SocketIO.Socket,
-  params: ParameterConfiguration[],
+  params: Param[],
   eventArgs
 ): any[] {
-  let args = [];
+  /**
+   * Final arguments
+   */
+  const args: any[] = [];
   /**
    * if no parameters provided, return io and socket, and event arguments (which came for default handler)
    */
@@ -47,29 +54,42 @@ function extractParameters(
 
   /**
    * Callback function
+   * @description extract callback function, it it exists
    * @type {Function}
    */
-  let callback: Function;
-  /**
-   * extract callback function, it it exists
-   */
-  if (eventArgs.length && typeof eventArgs[eventArgs.length - 1] === 'function') {
-    callback = eventArgs.pop();
-  }
+  const callback: Function =
+    eventArgs.length && typeof eventArgs[eventArgs.length - 1] === 'function' ?
+    eventArgs.pop() : noop;
 
   /**
    * loop through all params and put them into correct order
    */
   for (let item of params) {
     switch (item.type) {
-      default: args[item.index] = getSocket(item, socket); break; // socket
       case ParameterType.IO: args[item.index] = io; break;
+      case ParameterType.Socket: args[item.index] = getSocket(item, socket); break;
       case ParameterType.Args: args[item.index] = eventArgs.pop(); break;
-      case ParameterType.Callback: args[item.index] = callback || noop; break;
+      default: args[item.index] = callback; break;
     }
   }
 
   return args;
+}
+
+
+
+function makeEventListener(
+  controller,
+  listener,
+  io,
+  socket,
+  params
+) {
+  return (...args) => {
+    let handlerArgs = extractArguments(io, (socket || args.pop()), params[listener], args);
+
+    return controller[listener].apply(controller, handlerArgs);
+  };
 }
 
 /**
@@ -83,13 +103,16 @@ function extractParameters(
 function applyListeners(
   io: SocketIO.Server | SocketIO.Namespace,
   socket: SocketIO.Socket,
-  controller,
-  listeners: Listener,
-  params: Params
+  controller: SocketIOClass,
+  listeners: Listener[],
+  params: Param[]
 ) {
-  for (let listener of Object.keys(listeners)) {
+  for (const listener of Object.keys(listeners)) {
+
+    let actual = listeners[listener];
+
     if (socket) {
-      listeners[listener].middleware.forEach((middleware: Function) => {
+      actual.middleware.forEach((middleware: Function) => {
         (<any>socket).use((packet, next) => {
           if (packet[0] === listeners[listener].event) {
             return middleware(io, socket, packet, next);
@@ -99,42 +122,50 @@ function applyListeners(
       });
     }
 
-    (socket || io).on(listeners[listener].event, (...args) => {
-      let handlerArgs = extractParameters(io, (socket || args[0]), params[listener], args);
-      return controller[listener].apply(controller, handlerArgs);
-    });
+    (socket || io).on(actual.event, makeEventListener(controller, listener, io, socket, params));
   }
 }
 
 /**
  * Get artifacts, instantiates controller and extract meta data
- * @param Controller
- * @param deps
- * @returns { {controller, meta: SocketIOMeta, listeners: {io: Listener, socket: Listener}, params: Params} }
+ *
+ * @param {SocketIOClass} Controller
+ * @param {any[]} deps
+ * @returns {DecoratorsArtifacts}
  */
-function getArtifacts(Controller, deps) {
+function getArtifacts(Controller: SocketIOClass, deps: any[]): DecoratorsArtifacts {
   const controller = new Controller(...deps);
-  const meta: SocketIOMeta = controller.__meta__;
-  const namespace: string = meta.namespace;
-  const listeners = meta.listeners;
-  const params = meta.params;
+  const meta: Meta = getMeta(controller);
 
-  return { controller, meta, listeners, params, namespace };
+  return { controller, meta };
 }
 
 /**
  * Attach controller to socket
- * @param io
- * @param socket
- * @param artifacts
+ *
+ * @param {SocketIO.Server|SocketIO.Namespace} io
+ * @param {SocketIO.Socket} socket
+ * @param {DecoratorsArtifacts} artifacts
  */
-function _attachControllerToSocket(io, socket, artifacts) {
+function attachControllerToSocket(
+  io: SocketIO.Server|SocketIO.Namespace,
+  socket: SocketIO.Socket,
+  artifacts: DecoratorsArtifacts
+) {
   /**
    * Apply all registered middleware to socket
    */
   artifacts.meta.middleware.socket.forEach(middleware => {
     (<any>socket).use((...args) => middleware.apply(middleware, [io, socket, ...args]));
   });
+  artifacts.meta.middleware
+    .filter((md: Middleware) => md.type === MiddlewareType.Socket)
+    .reduce((acc: Function[], md: Middleware) => {
+      return acc.concat(md.middleware);
+    }, [])
+    .forEach((fn: Function) => {
+      io.use((...args) => fn.apply(fn, [io, ...args]));
+    });
 
   /**
    * Apply all registered controller-based middleware to socket
@@ -155,56 +186,81 @@ function _attachControllerToSocket(io, socket, artifacts) {
     io,
     socket,
     artifacts.controller,
-    artifacts.listeners.socket,
-    artifacts.params
+    artifacts.meta.listeners,
+    artifacts.meta.params
   );
 }
 
 /**
  * Attach Controller
- * @param io
- * @param Controller
+ *
+ * @param {SocketIO.Server} IO
+ * @param {SocketIOClass} Controller
+ * @param {any[]} deps
  */
-function attachController(io: SocketIO.Server, Controller, deps) {
+function attachController(
+  IO: SocketIO.Server,
+  Controller: SocketIOClass,
+  deps: any[]
+) {
   const artifacts = getArtifacts(Controller, deps);
-  const _io: SocketIO.Namespace = io.of(artifacts.namespace);
+  const io: SocketIO.Namespace = IO.of(artifacts.meta.ns || DEFAULT_NAMESPACE);
+  const socket: SocketIO.Socket = null;
 
   /**
-   * Apply all registered global middleware to io
+   * Filter all registered middleware and find IO middleware
+   * Construct flat array of functions
+   * Apply all middleware functions to io
    */
-  artifacts.meta.middleware.io.forEach(middleware => {
-    _io.use((...args) => middleware.apply(middleware, [_io, ...args]));
-  });
+  artifacts.meta.middleware
+    .filter((md: Middleware) => md.type === MiddlewareType.IO)
+    .reduce((acc: Function[], md: Middleware) => {
+      return acc.concat(md.middleware);
+    }, [])
+    .forEach((fn: Function) => {
+      io.use((...args) => fn.apply(fn, [io, ...args]));
+    });
 
   /**
    * Apply global listeners (io based)
    */
-  applyListeners(_io, null, artifacts.controller, artifacts.listeners.io, artifacts.params);
+  applyListeners(
+    io,
+    socket,
+    artifacts.controller,
+    artifacts.meta.listeners, artifacts.meta.params);
 
   /**
    * Apply local listeners (socket based)
    */
-  _io.on('connection', (socket) => _attachControllerToSocket(_io, socket, artifacts));
+  io.on('connection', (socket: SocketIO.Socket) => {
+    attachControllerToSocket(io, socket, artifacts);
+  });
 }
 
 /**
  * Attaches controllers to IO server
+ *
  * @param {SocketIO.Server} io
  * @param {Array<Injectable|ExpressClass>} injectables
  */
-export function attachControllers(io: SocketIO.Server, injectables: Array<Injectable | SocketIOClass>) {
-  injectables
-    .forEach((injectable: Injectable | SocketIOClass) => {
-      const controller = (<Injectable>injectable).provide || <SocketIOClass>injectable;
-      const deps = (<Injectable>injectable).deps || [];
+export function attachControllers(
+  io: SocketIO.Server,
+  injectables: any[]
+) {
+  injectables.forEach((injectable: Injectable | SocketIOClass) => {
+    const controller: SocketIOClass =
+      (<Injectable>injectable).provide || <SocketIOClass>injectable;
+    const deps: any[] = (<Injectable>injectable).deps || [];
 
-      attachController(io, controller, deps);
-    });
+    attachController(io, controller, deps);
+  });
 }
 
 /**
  * Attach Controller to already existed socket io server
  * With this approach you can't define global middleware, it's up to you.
+ *
  * @param {SocketIO.Server} io
  * @param {SocketIO.Socket} socket
  * @param {Array<Injectable|ExpressClass>} injectables
@@ -212,27 +268,13 @@ export function attachControllers(io: SocketIO.Server, injectables: Array<Inject
 export function attachControllersToSocket(
   io: SocketIO.Server,
   socket: SocketIO.Socket,
-  injectables: Array<Injectable | SocketIOClass>
+  injectables: any[]
 ) {
-  injectables
-    .forEach((injectable: Injectable | SocketIOClass) => {
-      const controller = (<Injectable>injectable).provide || <SocketIOClass>injectable;
-      const deps = (<Injectable>injectable).deps || [];
+  injectables.forEach((injectable: Injectable | SocketIOClass) => {
+    const controller: SocketIOClass =
+      (<Injectable>injectable).provide || <SocketIOClass>injectable;
+    const deps: any[] = (<Injectable>injectable).deps || [];
 
-      _attachControllerToSocket(io, socket, getArtifacts(controller, deps));
-    });
+    attachControllerToSocket(io, socket, getArtifacts(controller, deps));
+  });
 }
-
-/**
- * @alias
- * @see attachControllers
- * @deprecated
- */
-export let bootstrapSocketIO = attachControllers;
-
-/**
- * @alias
- * @see attachControllersToSocket
- * @deprecated
- */
-export let attachControllerToSocket = attachControllersToSocket;